修复了主页导航栏右侧有时无法及时刷新的bug
