+++ web/src/App.vue
@@ -9,6 +9,8 @@
 import NavBar from '@/components/NavBar.vue';
 import 'bootstrap/dist/css/bootstrap.min.css'
 import 'bootstrap/dist/js/bootstrap.min.js'
+import { useStore } from 'vuex'
+import router from './router';
 // import $ from 'jquery'
 
 export default {
@@ -17,6 +19,22 @@ export default {
     NavBar
   },
   setup() {
+    const store = useStore();
+    const jwt_token = localStorage.getItem("jwt_token");
+    if (jwt_token) {
+      store.commit("updateToken", jwt_token);
+      store.dispatch("getinfo", {
+        success() {
+          router.push({ name: 'Home' });
+          store.commit("updatePullingInfo", false);
+        },
+        error() {
+          store.commit("updatePullingInfo", false);
+        }
+      })
+    } else {
+      store.commit("updatePullingInfo", false);
+    }
     // let token = null;
     // $.ajax({
     //   url: "http://127.0.0.1:8088/user/account/token/",
