+++ web/src/views/user/account/LoginView.vue
@@ -40,21 +40,6 @@ export default {
         let password = ref('');
         let error_message = ref('');
 
-        const jwt_token = localStorage.getItem("jwt_token");
-        if (jwt_token) {
-            store.commit("updateToken", jwt_token);
-            store.dispatch("getinfo", {
-                success() {
-                    router.push({ name: "Home" });
-                    store.commit("updatePullingInfo", false);
-                },
-                error() {
-                    store.commit("updatePullingInfo", false);
-                }
-            })
-        } else {
-            store.commit("updatePullingInfo", false);
-        }
 
 
         const login = () => {
